缩减微服务的子模块数量 #9166 artifactory迁移子模块store
