+++ src/backend/ci/core/artifactory/api-artifactory/src/main/kotlin/com/tencent/devops/artifactory/api/ServiceArchiveAtomFileResource.kt
