+++ src/backend/ci/core/artifactory/api-artifactory/src/main/kotlin/com/tencent/devops/artifactory/pojo/ArchiveAtomRequest.kt
