+++ src/backend/ci/core/artifactory/biz-artifactory-sample/build.gradle.kts
@@ -27,5 +27,5 @@
 
 dependencies {
     api(project(":core:artifactory:api-artifactory-sample"))
-    api(project(":core:artifactory:biz-artifactory-store"))
+    api(project(":core:artifactory:biz-artifactory"))
 }
