+++ src/backend/ci/core/worker/worker-api-sdk/build.gradle.kts
@@ -28,7 +28,6 @@
 dependencies {
     api("com.google.code.gson:gson")
     api(project(":core:worker:worker-common"))
-    api(project(":core:artifactory:api-artifactory-store"))
     api("com.tencent.bk.repo:api-generic")
     api("com.tencent.bk.repo:api-repository")
 }
