+++ src/backend/ci/settings.gradle.kts
@@ -147,10 +147,8 @@ include(":core:environment:model-environment")
 include(":core:artifactory")
 include(":core:artifactory:api-artifactory")
 include(":core:artifactory:api-artifactory-sample")
-include(":core:artifactory:api-artifactory-store")
 include(":core:artifactory:biz-artifactory")
 include(":core:artifactory:biz-artifactory-sample")
-include(":core:artifactory:biz-artifactory-store")
 include(":core:artifactory:boot-artifactory")
 include(":core:artifactory:model-artifactory")
